requirements.txt
Adafruit-Blinka==8.10.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-dht==3.7.8
adafruit-circuitpython-requests==1.12.11
adafruit-circuitpython-typing==1.8.3
Adafruit-PlatformDetect==3.38.0
Adafruit-PureIO==1.1.9
board==1.0
colorzero==2.0
gpiozero==1.6.2
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
typing-extensions==4.4.0
